+++ crates/web-client/js/types/index.d.ts
@@ -29,6 +29,7 @@ export {
   NoteType,
   OutputNote,
   OutputNotesArray,
+  ProverWrapper,
   Rpo256,
   SerializedAccountHeader,
   TestUtils,