+++ crates/web-client/test/mocha.global.setup.mjs
@@ -80,6 +80,7 @@ before(async () => {
         OutputNote,
         OutputNotesArray,
         Rpo256,
+        ProverWrapper,
         TestUtils,
         TransactionFilter,
         TransactionRequest,
@@ -124,6 +125,7 @@ before(async () => {
       window.OutputNote = OutputNote;
       window.OutputNotesArray = OutputNotesArray;
       window.Rpo256 = Rpo256;
+      window.ProverWrapper = ProverWrapper;
       window.TestUtils = TestUtils;
       window.TransactionFilter = TransactionFilter;
       window.TransactionRequest = TransactionRequest;